rename 'buf.rs' to 'recv.rs'
